remove <Timer/> component entirely
