+++ modules/ccc-calendar/index.tsx
@@ -1,7 +1,7 @@
 import {EventList, type PoweredBy} from '@frogpond/event-list'
 import type {EventType} from '@frogpond/event-type'
 import {NoticeView} from '@frogpond/notice'
-import {useMomentTimer} from '@frogpond/timer/hook'
+import {useMomentTimer} from '@frogpond/timer'
 import {UseQueryResult} from '@tanstack/react-query'
 import * as React from 'react'
